LayoutTests/imported/w3c/web-platform-tests/cookies/samesite/window-open-reload.https-expected.txt

CONSOLE MESSAGE: [object MessageEvent]
Blocked access to external URL https://www1.localhost:9443/cookies/samesite/resources/puppet.html

Harness Error (TIMEOUT), message = null

PASS Reloaded same-host auxiliary navigations are strictly same-site.
TIMEOUT Reloaded subdomain auxiliary navigations are strictly same-site. Test timed out
NOTRUN Reloaded cross-site auxiliary navigations are laxly same-site
